+++ packages/shared/exceptions/index.ts
@@ -1,0 +1,3 @@
+export { ArgumentException } from './argument-exception';
+export { NotImplementedException } from './not-implemented-exception';
+export { NotSupportedException } from './not-supported-exception';